Reduce number of iterations
